+++ src/qbittorrentapi/_version_support.py
@@ -54,10 +54,11 @@ APP_VERSION_2_API_VERSION_MAP: dict[str, str] = {
     "v4.6.0rc1": "2.9.2",
     "v4.6.0rc2": "2.9.2",
     "v4.6.0": "2.9.2",
+    "v4.6.1": "2.9.3",
 }
 
-MOST_RECENT_SUPPORTED_APP_VERSION: Final[Literal["v4.6.0"]] = "v4.6.0"
-MOST_RECENT_SUPPORTED_API_VERSION: Final[Literal["2.9.2"]] = "2.9.2"
+MOST_RECENT_SUPPORTED_APP_VERSION: Final[Literal["v4.6.1"]] = "v4.6.1"
+MOST_RECENT_SUPPORTED_API_VERSION: Final[Literal["2.9.3"]] = "2.9.3"
 
 
 @lru_cache(maxsize=None)